requirements.txt
torch
torchaudio
librosa
numpy
scipy
matplotlib
pyyaml
